+++ src/GUISpinner.js
@@ -9,7 +9,7 @@
 		if(append == null || append == undefined) { append = true; }
 		
 		this.id = id;	
-		this.class = options.class || "GUISpinner";
+		this.className  = options.className  || "GUISpinner";
 		this.html = document.body || document.getElementsByTagName('body')[0];
 		this.numberPosition = {x:options.x, y:options.y};
 		this.numberSize = {width:options.w, height:options.h};	
@@ -49,7 +49,7 @@
 		number.style.position = "absolute";
 		number.id = this.id;	
 		number.name = this.id;
-		number.class = this.class;
+		number.className  = this.className;
 		number.tabindex = this.tabindex;
 		number.style.zIndex = this.zIndex;		
 		number.oninput = this.onchangeNumber;