+++ src/GUICheckbox.js
@@ -9,7 +9,6 @@
 		if(append == null || append == undefined) { append = true; }
 		
 		this.id = id;	
-		this.className  = options.className  || "GUICheckbox";
 		this.html = document.body || document.getElementsByTagName('body')[0];
 		this.checkboxPosition = {x:options.x, y:options.y};
 		this.checkboxSize = options.size || 1.0;
@@ -42,7 +41,7 @@
 		checkbox.style.margin = "0px";
 		checkbox.id = this.id;	
 		checkbox.name = this.id;		
-		checkbox.className  = this.className;
+		checkbox.className = "GUICheckbox";
 		checkbox.tabindex = this.tabindex;
 		checkbox.style.zIndex = this.zIndex;
 		checkbox.onclick = this.onClickCheckbox;