+++ src/GUIDialog.js
@@ -10,7 +10,6 @@
 				
 		this.id = id;
 		this.html = document.body || document.getElementsByTagName('body')[0];	
-		this.className  = options.className  || "GUIDialog";
 		this.dialogSize = {width:options.w, height:options.h};				
 		this.dialogPosition = {x:options.x, y:options.y};
 		this.overflow = options.overflow || "auto";
@@ -41,7 +40,7 @@
 		dialog.style.position = "absolute";
 		dialog.id = this.id;
 		dialog.name = this.id;
-		dialog.className  = this.className;
+		dialog.className = "GUIDialog";
 		dialog.style.zIndex = this.zIndex;
 		dialog.style.background = this.colorDialog;
 		dialog.style.borderRadius = this.radius+"px";