+++ src/GUIMeter.js
@@ -10,7 +10,6 @@
 		
 		this.meter = null
 		this.id = id;	
-		this.className  = options.className  || "GUIMeter";
 		this.html = document.body || document.getElementsByTagName('body')[0];
 		this.meterPosition = {x:options.x, y:options.y};
 		this.meterSize = {width:options.w, height:options.h};	 
@@ -50,7 +49,7 @@
 		this.meter.style.position = "absolute";
 		this.meter.id = this.id;	
 		this.meter.name = this.id;	
-		this.meter.className  = this.className;
+		this.meter.className = "GUIMeter";
 		this.meter.tabindex = this.tabindex;
 		this.meter.style.zIndex = this.zIndex;
 		if(this.orient == "vertical" || this.orient == "Vertical") {