+++ src/GUIPanel.js
@@ -9,7 +9,6 @@
 		if(append == null || append == undefined) { append = true; }
 		
 		this.id = id;	
-		this.className  = options.className  || "GUIMeter";
 		this.html = document.body || document.getElementsByTagName('body')[0];
 		this.divPosition = {x:options.x, y:options.y};
 		this.divSize = {width:options.w, height:options.h};
@@ -52,7 +51,7 @@
 		}		
 		div.id = this.id;	
 		div.name = this.id;
-		div.className  = this.className;		
+		div.className = "GUIPanel";		
 		div.style.position = this.position;
 		div.style.overflow = this.overflow;
 		div.style.zIndex = this.zIndex;