+++ src/GUIProgress.js
@@ -10,7 +10,6 @@
 		
 		this.progress = null;
 		this.id = id;	
-		this.className  = options.className  || "GUIProgress";
 		this.html = document.body || document.getElementsByTagName('body')[0];
 		this.progressPosition = {x:options.x, y:options.y};
 		this.progressSize = {width:options.w, height:options.h};	
@@ -50,7 +49,7 @@
 		}
 		this.progress.style.position = "absolute";
 		this.progress.id = this.id;
-		this.progress.className  = this.className;
+		this.progress.className = "GUIProgress";
 		this.progress.tabindex = this.tabindex;
 		this.progress.name = this.id;	
 		this.progress.style.zIndex = this.zIndex;